#17 Orders http://pm.5psolutions.net/public/index.php?path_info=projects/training-teodora-acatrinei/tasks/17 added total column in table

+++ database/migrations/2022_07_18_074723_create_products_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->id();
             $table->string('title');
             $table->text('description');
-            $table->float('price', 8, 2);
+            $table->decimal('price', 8, 2);
             $table->string('img');
             $table->softDeletes();
             $table->timestamps();
